# Sklep.Mobile/Sklep.Mobile/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sklep.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/ProductType": {
      "get": {
        "tags": [
          "ProductType"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProductType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProductType/{id}": {
      "get": {
        "tags": [
          "ProductType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProductType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ProductType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Order": {
        "required": [
          "city",
          "country",
          "email",
          "firstName",
          "houseNumber",
          "lastName",
          "phoneNumer",
          "postalCode",
          "province",
          "street"
        ],
        "type": "object",
        "properties": {
          "idOrder": {
            "type": "integer",
            "format": "int32"
          },
          "orderDate": {
            "type": "string",
            "format": "date-time"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "firstName": {
            "maxLength": 69,
            "type": "string"
          },
          "lastName": {
            "maxLength": 69,
            "type": "string"
          },
          "street": {
            "maxLength": 69,
            "type": "string"
          },
          "houseNumber": {
            "maxLength": 20,
            "type": "string"
          },
          "apartmentNumber": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 69,
            "type": "string"
          },
          "province": {
            "maxLength": 69,
            "type": "string"
          },
          "postalCode": {
            "maxLength": 10,
            "type": "string"
          },
          "country": {
            "maxLength": 69,
            "type": "string"
          },
          "phoneNumer": {
            "maxLength": 15,
            "type": "string"
          },
          "email": {
            "pattern": "[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,4}",
            "type": "string",
            "format": "email"
          },
          "orderItem": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "idOrderItem": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "idProduct": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "idOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          }
        },
        "additionalProperties": false
      },
      "Product": {
        "required": [
          "code",
          "created",
          "isActive",
          "modified",
          "price",
          "title"
        ],
        "type": "object",
        "properties": {
          "idProduct": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "picture": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "sale": {
            "type": "boolean"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "idProductType": {
            "type": "integer",
            "format": "int32"
          },
          "productType": {
            "$ref": "#/components/schemas/ProductType"
          },
          "idProductCategory": {
            "type": "integer",
            "format": "int32"
          },
          "productCategory": {
            "$ref": "#/components/schemas/ProductCategory"
          },
          "idProductProducer": {
            "type": "integer",
            "format": "int32"
          },
          "productProducer": {
            "$ref": "#/components/schemas/ProductProducer"
          },
          "orderItem": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductCategory": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "idProductCategory": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "product": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductProducer": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "idProductProducer": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "string",
            "format": "date-time"
          },
          "isActicve": {
            "type": "boolean"
          },
          "product": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductType": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "idProductType": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "product": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}